Fix regex exponential backtracking.
/(a|b)*c/ is exponential for runs of substrings that match both a and b.

In this case, the pattern /(\\.|(?!a).)/, that matches backslashed 'a',
is exponential in the number of backslash characters.

As a simple workaround, replace . with [^\\], thus making the two
branches non-overlapping. Note that the \\. must then be \\.? to still
allow non-escaped backslash characters.

(cherry picked from xbbcode 05107eaa3347f773b9138590b3bbd17f28b0939a)